requirements.txt
MySQL-python
django>=1.4
python-social-auth
django-compressor
jsmin
fabric
django-suit
django-reversion
django-pagedown
django-guardian
django-object-actions
django-taggit
django-autocomplete-light